# apps/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: PhyloPic API
  description: Documentation for the next version of *PhyloPic*'s Application Programming Interface.
  contact:
    name: Mike Keesey
    email: keesey+phylopic@gmail.com
    url: http://tmkeesey.net
  license:
    name: CC0
    url: https://creativecommons.org/share-your-work/public-domain/cc0
  version: 0.0.0
...
servers:
  - url: https://api.phylopic.org
    description: The production API
components:
  headers:
    Accept-Ranges:
      description: The units acceptable for range requests. Always `items`.
      example: items
      schema:
        pattern: "^items$"
        type: string
    Access-Control-Allow-Credentials:
      description: Specifies that credentials are allowed in requests. Always `true`.
      example: "true"
      schema:
        pattern: "^true$"
        type: string
    Access-Control-Allow-Origin:
      description: Specifies that requests may be made from any origin. Always `*`.
      example: "*"
      schema:
        pattern: '^\*$'
        type: string
    Cache-Control:
      description: Specifications for caching mechanisms.
      example: no-cache
      schema:
        type: string
    Content-Length:
      description: The length of the content, in bytes.
      example: 1024
      schema:
        format: number
        type: string
    Content-Range:
      description: The range of embedded items, in the format `"items {start}-{end}/{total}"`. If the total is unknown, it will be `"*"`. If there are no items, there will be `"*"` instead of `"{start}-{end}"`.
      example: items 0-9/88
      schema:
        pattern: '^items \d+-\d+/\d+$'
        type: string
    Content-Type:
      description: The media type for the payload, including API version. Always `application/vnd.phylopic.v2+json`.
      example: application/vnd.phylopic.v2+json
      schema:
        format: mimetype
        type: string
...
    ETag-GET-Range:
      description: An identifier for this version of the list. Should be used on subsequent calls to ensure the same version of the list is being loaded (`If-Match` header).
      example: '"identifier"'
      schema:
        pattern: '^"[^"]+"$'
        type: string
...
    Warning-GET-Range:
      description: Warning message. This will be included if you request a range starting after `0` and do not provide an `If-Match` header. Send back a header that matches the `ETag` header of the response to remove this warning.
      schema:
        type: string
  parameters:
    Accept:
      name: Accept
      in: header
      description: |
        The media type for the response, including API version. Must allow `application/vnd.phylopic.v2+json`.

        If omitted, `*/*` (a wildcard allowing all types) is assumed, and the response's `Content-Type` header will be `application/vnd.phylopic.v2+json`.
      schema:
        format: mimetype
        type: string
...
    If-Match-GET:
      name: If-Match
      in: header
      description: Identifies which version(s) of the list is/are expected, as given by the `ETag` header on a previous response. If that version is out of date, a `416` response will be returned.
      schema:
        pattern: '^(\*|(W\\/)?"([^"]|\\")+"(,\s*(W\/)?"([^"]|\\")+")*)$'
        type: string
      example: '"identifier"'
...
    Range:
      name: Range
      in: header
      required: true
      description: |
        A range of the items to read, inclusive.

        Example: `items=0-9`
      schema:
        pattern: '^items=\d+-\d+$'
        type: string
...
    embed-Node:
      name: embed
      in: query
      description: |
        A space-separated list of related entities to embed. May include:
        - `childNodes`
        - `parentNode`
        - `primaryImage`

        If omitted, no entities will be embedded.
      schema:
        pattern: "^((childNodes|parentNode|primaryImage) )*(childNodes|parentNode|primaryImage)$"
        type: string
...
    uuid-Entity:
      name: uuid
      in: path
      required: true
      description: The UUID of the entity.
      schema:
        format: uuid
        type: string
  responses:
...
    "400":
      description: There is an error in the request.
      headers:
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Content-Length:
          $ref: "#/components/headers/Content-Length"
        Content-Type:
          $ref: "#/components/headers/Content-Type"
      content:
        application/vnd.phylopic.v2+json:
          schema:
            $ref: "#/components/schemas/Errors"
    "404":
      description: The resource cannot be found.
      headers:
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Cache-Control:
          $ref: "#/components/headers/Cache-Control"
        Content-Length:
          $ref: "#/components/headers/Content-Length"
        Content-Type:
          $ref: "#/components/headers/Content-Type"
      content:
        application/vnd.phylopic.v2+json:
          schema:
            $ref: "#/components/schemas/Errors"
    "406":
      description: The request cannot be accepted because the `Accept` header does not allow an available type of content.
      headers:
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Content-Length:
          $ref: "#/components/headers/Content-Length"
        Content-Type:
          $ref: "#/components/headers/Content-Type"
      content:
        application/vnd.phylopic.v2+json:
          schema:
            $ref: "#/components/schemas/Errors"
    "410":
      description: The resource has been removed.
      headers:
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Cache-Control:
          $ref: "#/components/headers/Cache-Control"
        Content-Length:
          $ref: "#/components/headers/Content-Length"
        Content-Type:
          $ref: "#/components/headers/Content-Type"
      content:
        application/vnd.phylopic.v2+json:
          schema:
            $ref: "#/components/schemas/Errors"
...
    "416":
      description: The requested range is not satisfiable.
      headers:
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Content-Length:
          $ref: "#/components/headers/Content-Length"
        Content-Type:
          $ref: "#/components/headers/Content-Type"
        ETag:
          $ref: "#/components/headers/ETag-GET-Range"
      content:
        application/vnd.phylopic.v2+json:
          schema:
            $ref: "#/components/schemas/Errors"
...
    "list-Node":
      description: |
        [HAL](http://stateless.co/hal_specification.html) [JSON](https://www.json.org/) list of embedded node records.
      headers:
        Accept-Ranges:
          $ref: "#/components/headers/Accept-Ranges"
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Cache-Control:
          $ref: "#/components/headers/Cache-Control"
        Content-Length:
          $ref: "#/components/headers/Content-Length"
        Content-Range:
          $ref: "#/components/headers/Content-Range"
        Content-Type:
          $ref: "#/components/headers/Content-Type"
        ETag:
          $ref: "#/components/headers/ETag-GET-Range"
        Warning:
          $ref: "#/components/headers/Warning-GET-Range"
      content:
        application/vnd.phylopic.v2+json:
          schema:
            type: object
            required:
              - _embedded
              - _links
            properties:
              _embedded:
                type: object
                description: Container for embedded entities.
                properties:
                  items:
                    type: array
                    description: The requested node entities.
                    items:
                      $ref: "#/components/schemas/NodeWithEmbedded"
              _links:
                description: Container for a link to this resource (`self`).
                type: object
                required:
                  - self
                properties:
                  self:
                    allOf:
                      - $ref: "#/components/schemas/Link"
                      - description: Link to this list.
              total:
                description: The total number of items in the set.
                format: uint
                type: number
  schemas:
    Dated:
      description: Common properties for all dated objects.
      required:
        - created
        - modified
      type: object
      properties:
        created:
          description: The time this was created.
          format: date-time
          type: string
        modified:
          description: The last time this was modified.
          format: date-time
          type: string
    Entity:
      description: Common properties for all entities.
      required:
        - _links
        - created
        - modified
        - uuid
      allOf:
        - $ref: "#/components/schemas/Dated"
        - $ref: "#/components/schemas/EntityReference"
    EntityReference:
      description: Identifier and link for an entity.
      type: object
      required:
        - _links
        - uuid
      properties:
        _links:
          type: object
          required:
            - self
          properties:
            self:
              $ref: "#/components/schemas/Link"
        uuid:
          description: The universally unique identifier for this entity.
          format: uuid
          type: string
          example: 123e4567-e89b-12d3-a456-426655440000
    Errors:
      description: List of errors.
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            type: object
            required:
              - developerMessage
              - type
              - userMessage
            properties:
              developerMessage:
                description: A message with details for developers.
                type: string
              documentation:
                description: A link to documentation for the error.
                format: url
                type: string
              field:
                description: The path to the field in the payload that caused the error.
                type: string
                example: "attribution"
              type:
                description: The type of error, from [this list](https://docs.aws.amazon.com/apigateway/api-reference/resource/gateway-response/#responseType).
                type: string
                example: BAD_REQUEST_PARAMETERS
              userMessage:
                description: A message to be presented to the user.
                type: string
        stack:
          description: Details about which functions were being processed when the error occurred.
          type: string
    Image:
      allOf:
        - $ref: "#/components/schemas/Entity"
        - description: An image illustrating one or more nodes.
          required:
            - _links
            - attribution
            - sponsor
          type: object
          properties:
            attribution:
              type: string
              example: "Charles Knight"
              nullable: true
            sponsor:
              type: string
              example: "Patron Name"
              nullable: true
            _links:
              $ref: "#/components/schemas/ImageLinks"
    ImageLinks:
      type: object
      required:
        - contributor
        - generalNode
        - http://ogp.me/ns#image
        - license
        - nodes
        - rasterFiles
        - self
        - sourceFile
        - specificNode
        - thumbnailFiles
        - twitter:image
        - vectorFile
      properties:
        contributor:
          allOf:
            - $ref: "#/components/schemas/Link"
            - description: A `mailto:` URL with the contributor's email address.
        generalNode:
          allOf:
            - $ref: "#/components/schemas/Link"
            - description: A link to the node indicating how generally the image can be applied, either an ancestor of `specificNode` or a node indicating a common ancestor with `specificNode`.
              nullable: true
        "http://ogp.me/ns#image":
          allOf:
            - $ref: "#/components/schemas/MediaLink"
            - description: A link to an image for use with Open Graph (PNG format).
        license:
          allOf:
            - $ref: "#/components/schemas/Link"
            - description: A link to the license that this image is under.
        nodes:
          type: array
          description: Links to all nodes that this image illustrates, based on `specificNode` and (if provided) `generalNode`.
          items:
            $ref: "#/components/schemas/Link"
        rasterFiles:
          type: array
          description: A list of links to raster image files (PNG format) of various sizes.
          items:
            $ref: "#/components/schemas/MediaLink"
        sourceFile:
          allOf:
            - $ref: "#/components/schemas/MediaLink"
            - description: A link to the raster or vector image file as it was submitted. May be an SVG file or any of the accepted raster image formats.
        specificNode:
          allOf:
            - $ref: "#/components/schemas/Link"
            - description: A link to the node indicating the specific taxon that this image illustrates.
        submission:
          allOf:
            - $ref: "#/components/schemas/Link"
            - description: A link for viewing the latest submitted changes.
        thumbnailFiles:
          type: array
          description: A list of links to raster thumbnail image files (PNG format) of various square sizes.
          items:
            $ref: "#/components/schemas/MediaLink"
        "twitter:image":
          allOf:
            - $ref: "#/components/schemas/MediaLink"
            - description: A link to an image for use with Twitter (PNG format).
        vectorFile:
          allOf:
            - $ref: "#/components/schemas/MediaLink"
            - description: A link to the vector image file as it was submitted. Always in SVG format.
              nullable: true
...
    Link:
      description: A basic link.
      type: object
      required:
        - href
      properties:
        href:
          format: url
          type: string
          example: /path
    MediaLink:
      allOf:
        - $ref: "#/components/schemas/Link"
        - type: object
          description: A link to an image file.
          required:
            - href
            - sizes
            - type
          properties:
            href:
              example: "/path/123e4567-e89b-12d3-a456-426655440000/raster32.png"
            sizes:
              pattern: '^\d+x\d+$'
              type: string
              example: 32x32
            type:
              format: mimetype
              type: string
              example: "image/png"
    Node:
      allOf:
        - $ref: "#/components/schemas/Entity"
        - description: A phylogenetic node, corresponding to a taxonomic unit.
          type: object
          required:
            - _links
            - names
            - root
          properties:
            _links:
              type: object
              required:
                - childNodes
                - cladeImages
                - contributor
                - external
                - images
                - parentNode
                - primaryImage
                - self
              properties:
                childNodes:
                  type: array
                  description: |
                    Links to all immediate descendant nodes, in alphabetical order.
                    Note that this only includes illustrated nodes.
                  items:
                    allOf:
                      - $ref: "#/components/schemas/Link"
                cladeImages:
                  allOf:
                    - $ref: "#/components/schemas/Link"
                    - description: A link to a list of all images illustrating this node or any descendant node.
                contributor:
                  allOf:
                    - $ref: "#/components/schemas/Link"
                    - description: A `mailto:` URL with the contributor's email address.
                external:
                  type: array
                  description: Links to the same node under other domains (for example, Encyclopedia of Life).
                  items:
                    $ref: "#/components/schemas/TitledLink"
                images:
                  allOf:
                    - $ref: "#/components/schemas/Link"
                    - description: A link to a list of all images illustrating this node.
                parentNode:
                  allOf:
                    - $ref: "#/components/schemas/Link"
                    - description: Link to the parent node (immediate ancestor).
                      nullable: true
                primaryImage:
                  allOf:
                    - $ref: "#/components/schemas/Link"
                    - description: Link to the primary image for this node.
                      nullable: true
            names:
              type: array
              description: A list of all of this node's names, starting with the preferred name.
              items:
                $ref: "#/components/schemas/NodeName"
            root:
              description: A flag that, if `true`, indicates that this is a root node and should not have a parent.
              type: boolean
    NodeName:
      description: A taxonomic name, associated with a phylogenetic node.
      type: array
      items:
        type: object
        required:
          - class
          - text
        properties:
          class:
            format: "^(citation|comment|operator|rank|scientific|vernacular)$"
            type: string
            example: scientific
          text:
            format: '^\S+( \S+)*$'
            type: string
            example: Homo sapiens
...
    NodeWithEmbedded:
      description: A resource for a phylogenetic node.
      required:
        - _embedded
        - _links
        - names
        - root
      allOf:
        - type: object
          properties:
            _embedded:
              type: object
              description: Embedded entities related to the phylogenetic node. See `_links` for more information.
              properties:
                childNodes:
                  type: array
                  items:
                    $ref: "#/components/schemas/Node"
                contributor:
                  allOf:
                    - $ref: "#/components/schemas/Link"
                    - description: A `mailto:` URL with the contributor's email address.
                parentNode:
                  allOf:
                    - $ref: "#/components/schemas/Node"
                    - nullable: true
                primaryImage:
                  allOf:
                    - $ref: "#/components/schemas/Image"
                    - nullable: true
        - $ref: "#/components/schemas/Node"
    TitledLink:
      description: A link with a title, usually for an external resource.
      allOf:
        - $ref: "#/components/schemas/Link"
        - type: object
          required:
            - title
          properties:
            title:
              type: string
              example: "Name"
  securitySchemes:
    keyAuth:
      description: User consuming the API data.
      type: apiKey
      in: header
      name: X-API-Key
paths:
...
  /lineage/{uuid}:
    get:
      operationId: getLineage
      summary: Retrieves the ancestral lineage for a given node.
      description: The nodes are listed in order from least to most ancestral. That is, the first node is the specified node, the second is its parent node, and the last is the root node.
      tags:
        - nodes
      parameters:
        - $ref: "#/components/parameters/Accept"
        - $ref: "#/components/parameters/If-Match-GET"
        - $ref: "#/components/parameters/Range"
        - $ref: "#/components/parameters/embed-Node"
        - $ref: "#/components/parameters/uuid-Entity"
      security:
        - keyAuth: []
      responses:
        "206":
          $ref: "#/components/responses/list-Node"
        "400":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
        "406":
          $ref: "#/components/responses/406"
        "410":
          $ref: "#/components/responses/410"
        "416":
          $ref: "#/components/responses/416"
...